# openapi.yaml
openapi: 3.0.3
info:
  title: Coding Interview Platform API
  description: |
    Backend API for the Coding Interview Platform. This API supports:
    - User authentication (login, signup, logout)
    - Interview session management
    - Real-time code collaboration
    - Question management
    - Chat/messaging
    - Private notes
    - Permission controls
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.interview-platform.com/v1
    description: Production server
...
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: Authenticate a user with username and return user data with auth token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
              properties:
                username:
                  type: string
                  example: "john_doe"
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Create a new user account with specified role
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - role
              properties:
                username:
                  type: string
                  example: "john_doe"
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                role:
                  $ref: '#/components/schemas/UserRole'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Invalidate user session
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Successfully logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the currently authenticated user's information
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # Session Endpoints
  # ============================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    UserRole:
      type: string
      enum:
        - interviewer
        - candidate
      example: "interviewer"
...
    User:
      type: object
      required:
        - id
        - username
        - email
        - role
      properties:
        id:
          type: string
          example: "usr_abc123"
        username:
          type: string
          example: "john_doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          $ref: '#/components/schemas/UserRole'
    AuthResponse:
      type: object
      required:
        - user
        - token
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT authentication token
          example: "eyJhbGciOiJIUzI1NiIs..."
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "NOT_FOUND"